refactor(backend): split reference/service.py by domain
Decomposes the 722-LOC reference/service.py into a sub-package at
the same import path:

  reference/service/
    __init__.py        (75)   re-exports everything consumers use
    types.py           (42)   SyncResult, ReferenceDataStatus dataclasses
    constants.py       (70)   HNF1B_DOMAINS, HNF1B_EXONS, Ensembl URLs
    hnf1b_importer.py  (299)  initialize_reference_data + HNF1B chain helpers
    ensembl_sync.py    (180)  fetch + parse + sync_chr17q12_genes
    status.py          (68)   get_reference_data_status query

Every file is under 500 LOC. Public import path unchanged, so the
admin endpoints (status_routes, sync_service, sync_reference_routes)
continue to work without modification.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
